+++ straxen/plugins/raw_records_sv/__init__.py
@@ -1,2 +1,5 @@
-from . import raw_records_sv
-from .raw_records_sv import *
+# from . import raw_records_sv
+# from .raw_records_sv import *
+
+from . import radial_veto
+from .radial_veto import *
